requirements.txt
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.3.0
mlflow>=2.7.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0